Cleaning up data pt.2

+++ app/src/main/java/eh/workout/journal/com/workoutjournal/util/EquationsHelper.java
@@ -10,7 +10,7 @@ public class EquationsHelper {
     public static double getOneRepMax(String w, String r) {
         double weight = Double.valueOf(w);
         double reps = Double.valueOf(r);
-        return (double) weight / (1.0278 - (0.0278 * reps));
+        return weight / (1.0278 - (0.0278 * reps));
     }
 
     public static int getOneRepMaxInt(double orm) {
